create a new genericlinux/buses package, doesn't compile yet
